Refactor and introduce classifiers
